requirements.txt
python-dotenv
pdfkit
numpy
pandas
requests
openai
chromadb
langchain
llama_index

